Update UserKindController and TradeController

+++ realestate/src/main/java/vn/edu/uit/realestate/Controller/UserKindController.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,6 +21,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import vn.edu.uit.realestate.Controller.ExceptionHandler.NotFoundException;
+import vn.edu.uit.realestate.Model.Trade;
 import vn.edu.uit.realestate.Model.User;
 import vn.edu.uit.realestate.Model.UserKind;
 import vn.edu.uit.realestate.Repository.UserKindRepository;
@@ -29,55 +32,44 @@ import vn.edu.uit.realestate.Repository.UserRepository;
 public class UserKindController {
 	@Autowired
 	private UserKindRepository userKindRepository;
-	private UserRepository userRepository;
 
     @GetMapping("/userkinds")
-    public List<UserKind> getUserKinds() {
+    public ResponseEntity<List<UserKind>> getUserKinds() {
     	List<UserKind> userKinds = userKindRepository.findAll();
-        return userKinds;
+    	if (userKinds.isEmpty() == true) {
+    		throw new NotFoundException("Cannot find any User Kind");
+    	}
+        return new ResponseEntity<>(userKinds,HttpStatus.OK);
     }
-    @GetMapping("/userkind/{id}")
-    public Optional<UserKind> getUserKind(@PathVariable long id) {
-		try {
-		    	Optional<UserKind> userKind = userKindRepository.findById(id);
-		        return userKind;
-		}catch(Exception e) {
-			System.out.println("\nPROBLEM: "+e);
-		}
-		return null;
+    @GetMapping("/userkinds/{id}")
+    public ResponseEntity<UserKind> getUserKind(@PathVariable long id) {
+    	Optional<UserKind> foundUserKind = userKindRepository.findById(id);
+		if (foundUserKind.isPresent()==false) {
+    		throw new NotFoundException("Cannot find any User Kind with id="+id);
+    	}
+        return new ResponseEntity<>(foundUserKind.get(), HttpStatus.OK);
     }
-//    @GetMapping("/userkind/{userKindId}/users")
-//    public Page<User> getUsersByUserKindId(@PathVariable Long userKindId) {
-//    	Page<User> users = userRepository.findByUserKindId(userKindId, PageRequest.of(0, 5));
-//    	if(users == null)
-//    		return null;
-//        return users;
-//    }
-//    @GetMapping("/userkind/{userKindId}/users/{userId}")
-//    public Optional<User> getUsersByIdAndUserKindId(@PathVariable (value = "userKindId") Long userKindId,
-//    		@PathVariable (value = "userId") Long userId,
-//            Pageable pageable) {
-//    	Optional<User> users = userRepository.findByIdAndUserKindId(userId, userKindId);
-//        return users;
-//    }
+    @GetMapping("/userkinds/{userKindId}/users")
+    public ResponseEntity<List<User>> getUsersByUserKindId(@PathVariable Long userKindId) {
+    	UserKind foundUserKind = userKindRepository.findById(userKindId).get();
+    	List<User> users = foundUserKind.getUsers();
+    	if(users == null)
+    		throw new NotFoundException("Cannot find any User with User Kind Id="+userKindId);
+        return new ResponseEntity<>(users, HttpStatus.OK);
+    }
     @PostMapping("/userkinds")
-    public ResponseEntity<UserKind> postUserKind(@Valid @RequestBody UserKind userKind) {
+    public ResponseEntity<UserKind> postUserKind(@RequestBody UserKind userKind) {
     	userKindRepository.save(userKind);
     	URI location = ServletUriComponentsBuilder
     			.fromCurrentRequest().path("/{id}")
     			.buildAndExpand(userKind.getId()).toUri();
-    	///ResponseEntity return (status + header + body)
-    	///Example: var headers = new HttpHeaders();
-    	///headers.add("Responded", "MyController");
-    	///ResponseEntity.accepted().headers(headers).body(c);
     	return ResponseEntity.created(location).build();
     }
-    @DeleteMapping("/userkind/{id}")
+    @DeleteMapping("/userkinds/{id}")
     public void deleteUserKind(@PathVariable long id) {
-		try {
-		    	userKindRepository.deleteById(id);
-		}catch(Exception e) {
-			System.out.println("\nPROBLEM: "+e);
+    	if(!userKindRepository.existsById(id)) {
+			throw new NotFoundException("Cannot find any User Kind with Id="+id);
 		}
+    	userKindRepository.deleteById(id);
     }
 }
